comzey_api/requirements.txt
asgiref==3.4.1 
django==3.2.6 
pytz==2021.1 
sqlparse==0.4.1
djangorestframework==3.12.4
django-filter==2.4.0
psycopg2-binary==2.9.1
python-decouple==3.4
djangorestframework-jwt==1.11.0
django-cors-headers==3.8.0
django-location-field
requests
pyfcm
python-dateutil
mailjet_rest